2022/07_observer_metamerism/requirements.txt
opencv-python
cycler
numpy
matplotlib
scipy
colour-science
colour-datasets
pillow
sphinx_rtd_theme
sympy
flake8
pytest
jupyter
ipywidgets
pysimplegui
jupyterlab
wavio
pyside2
pyqtgraph
